@@ -5,19 +5,26 @@
   "keywords": [],
   "main": "src/index.js",
   "dependencies": {
-    "@testing-library/jest-dom": "5.16.1",
-    "@testing-library/react": "12.1.2",
-    "@testing-library/user-event": "13.5.0",
-    "react": "17.0.0",
-    "react-dom": "17.0.0",
-    "react-scripts": "3.4.3",
-    "typescript": "4.5.4"
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "13.4.0",
+    "@testing-library/user-event": "14.4.3",
+    "@types/react": "^18.0.24",
+    "@types/react-dom": "^18.0.8",
+    "@types/uuid": "^8.3.4",
+    "react": "^18.2.0",
+    "react-dom": "18.2.0",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.55.0",
+    "sass-loader": "^13",
+    "typescript": "4.8.4",
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "@types/react": "^18.0.18",
-    "@types/react-dom": "^18.0.6",
-    "node-sass": "^4.0.0",
-    "typescript": "3.8.3"
+    "@types/react": "^18.0.24",
+    "@types/react-dom": "^18.0.8",
+    "sass": "^1.55.0",
+    "sass-loader": "^13",
+    "typescript": "4.8.4"
   },
   "scripts": {
     "start": "react-scripts start",
